Skip whitelist checks during a force unblock.
When a block record is being force unblocked and a save() is called on
the Block object, skip any whitelist checks.

This fixes the problem caused by

 * blocking an ip
 * whitelisting that ip
 * trying to unblock that ip

During the final save() the whitelist check was preventing the
Block object from being saved.  If unblock_now is set, any whitelist
checks can be skipped.

Whitelist checks really only need to be ran on a save if block.cidr
changed, but that will require a more involved code change.

This fixes #3