Handle unicode characters in log messages better
This was breaking two things:

urllib.quote can't handle unicode (seriously?)

The CSV is being output in ascii (which is probably for the best, so we
can't shove unicode into it).  This should possibly encode to utf-8, but
I have seen other libraries crash and burn when given a CSV file encoded
as utf-8.