problem: incorrect block queue query can cause SQL errors.
This fixes a bug introduced in 57c3d3b54b50484d366d4546c3673f1a6e4d05c6
that is also related to the fix in 729c6605ad0d2558e2a7a9bcd5476f50a258a53f
that attempts to fix problems caused by overlapping blocks.

The over aggressive 'be.removed is NULL' causes the 'fixed' blockentry
records to not match causing the block queue to return entries that are
already blocked.

This change hurts performance a bit, but due to the recent changes in
the API, it will only affect the first block queue query when a
bhr-client starts up.

The work outlined in #2 about clearing old bhr_blockentries should help
with this.